backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic-settings
sqladmin
python-multipart
psycopg2-binary==2.9.9
python-dotenv==1.0.0